+++ src/main/java/sk/stuba/fei/uim/vsa/pr2/web/controllers/CarTypeController.java
@@ -2,10 +2,13 @@ package sk.stuba.fei.uim.vsa.pr2.web.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.HttpHeaders;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import sk.stuba.fei.uim.vsa.pr2.domain.CarType;
+import sk.stuba.fei.uim.vsa.pr2.domain.User;
 import sk.stuba.fei.uim.vsa.pr2.web.controllers.service.CarTypeService;
+import sk.stuba.fei.uim.vsa.pr2.web.controllers.service.ServiceController;
 import sk.stuba.fei.uim.vsa.pr2.web.response.CarTypeDto;
 import sk.stuba.fei.uim.vsa.pr2.web.response.factories.CarTypeFactory;
 
@@ -22,7 +25,9 @@ public class CarTypeController {
     @GET
     @Path("/cartypes")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getCarTypes() {
+    public Response getCarTypes(@HeaderParam(HttpHeaders.AUTHORIZATION) String authorization) {
+        User userAuth = new ServiceController().getUserAuth(authorization);
+        if (userAuth == null) return Response.status(Response.Status.UNAUTHORIZED).build();
         // TODO: PARAMETER SEARCH
         List<CarType> carTypes = service.getCarTypes();
         if (carTypes.isEmpty()) {
@@ -36,7 +41,9 @@ public class CarTypeController {
     @GET
     @Path("/cartypes/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getCarType(@PathParam("id") Long id) {
+    public Response getCarType(@HeaderParam(HttpHeaders.AUTHORIZATION) String authorization, @PathParam("id") Long id) {
+        User userAuth = new ServiceController().getUserAuth(authorization);
+        if (userAuth == null) return Response.status(Response.Status.UNAUTHORIZED).build();
         CarType carType = service.getCarType(id);
         if (carType == null) {
             return Response.status(Response.Status.NOT_FOUND).build();
@@ -50,7 +57,9 @@ public class CarTypeController {
     @DELETE
     @Path("/cartypes/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response deleteCarType(@PathParam("id") Long id){
+    public Response deleteCarType(@HeaderParam(HttpHeaders.AUTHORIZATION) String authorization, @PathParam("id") Long id){
+        User userAuth = new ServiceController().getUserAuth(authorization);
+        if (userAuth == null) return Response.status(Response.Status.UNAUTHORIZED).build();
         CarType carType = service.deleteCarType(id);
         if(carType == null){
             return Response.status(Response.Status.NOT_FOUND).build();
